Bug fix for large volumes
git-svn-id: file:///home/roberto/Git/svn@9901 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: f1dd4a768d8c30e823cc9f43604993f639e6d3ed

+++ trunk/xmipp/libraries/data/multidim_array.h
@@ -1051,7 +1051,7 @@ public:
       */
     void setDimensions(int Xdim, int Ydim, int Zdim, int Ndim)
     {
-        if (Xdim*Ydim*Zdim*Ndim < 1)
+        if (((size_t)Xdim)*Ydim*Zdim*Ndim < 1)
             REPORT_ERROR(ERR_MULTIDIM_SIZE, "Dimensions' size cannot be zero nor negative.");
         ndim=Ndim;
         zdim=Zdim;
